Added density control

+++ src/main/java/com/fantasticsource/customparticles/client/factory/CustomParticleFactory.java
@@ -104,7 +104,7 @@ public abstract class CustomParticleFactory
         PathedParticle particle = create(x, y, z);
 
 
-        if (emitter.cullDistanceSquared != -1) particle.cullDistanceSquared = emitter.cullDistanceSquared;
+        if (emitter.cullDistanceSquared != -1) particle.cullDistanceSquared = (int) emitter.cullDistanceSquared;
 
         particle.spriteMetaData = spriteMetaData;
         particle.useFoliageColor = useFoliageColor;
